two separate components, pfBend1 halfway done
